# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CreatePaymentRequest:
    properties:
      planId:
        type: integer
    required:
    - planId
    type: object
...
  dto.PaymentResponse:
    properties:
      amount:
        type: integer
      createdAt:
        type: string
      currency:
        type: string
      id:
        type: integer
      paidAt:
        type: string
      paymentUrl:
        type: string
      planId:
        type: integer
      provider:
        $ref: '#/definitions/models.PaymentProvider'
      providerPaymentId:
        type: string
      status:
        $ref: '#/definitions/models.PaymentStatus'
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
...
  dto.SubscriptionPlanResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      durationDays:
        type: integer
      id:
        type: integer
      isActive:
        type: boolean
      name:
        type: string
      price:
        type: integer
      updatedAt:
        type: string
    type: object
  dto.SubscriptionResponse:
    properties:
      activeUntil:
        type: string
      cancelledAt:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      plan:
        $ref: '#/definitions/dto.SubscriptionPlanResponse'
      planId:
        type: integer
      status:
        $ref: '#/definitions/models.SubscriptionStatus'
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
...
  models.PaymentProvider:
    enum:
    - mock
    - kaspi
    - stripe
    type: string
    x-enum-varnames:
    - PaymentProviderMock
    - PaymentProviderKaspi
    - PaymentProviderStripe
  models.PaymentStatus:
    enum:
    - pending
    - paid
    - failed
    - expired
    type: string
    x-enum-varnames:
    - PaymentStatusPending
    - PaymentStatusPaid
    - PaymentStatusFailed
    - PaymentStatusExpired
...
  models.SubscriptionStatus:
    enum:
    - active
    - expired
    - cancelled
    type: string
    x-enum-varnames:
    - SubscriptionStatusActive
    - SubscriptionStatusExpired
    - SubscriptionStatusCancelled
...
host: localhost:8080
info:
  contact: {}
  description: Api server for DaraTil PLatform
  title: DaraTil Backend
  version: "0.2"
paths:
...
  /payments:
    post:
      consumes:
      - application/json
      description: Create a new payment for subscription plan. Returns payment URL
        for completing payment.
      parameters:
      - description: Payment create request
        in: body
        name: paymentReq
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create payment
      tags:
      - Payments
  /payments/{id}/mock-pay:
    get:
      consumes:
      - application/json
      description: Confirm mock payment by payment ID and activate user subscription.
        This endpoint is used only for local/testing payment flow.
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mock payment confirmation
      tags:
      - Payments
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"